XIVY-3372 Allow unmanaged users in a managed security system
Allow all modifications for users even the user is managed, like:
- Edit user infos
- Add/remove roles
- Add/remove user properties

If it's managed, the user will be updated the next time.

Streamline the name of something which is managed by the external
security system. Use the term 'manage' instead of 'backed' or 'synced'.

Show a small icon to see which components are managed by the
external security system.