XIVY-11213 Add a new Flight Recorder View to Engine Cockpit
Add a new Flight Recorder view to the Engine Cockpit that allows to
start, stop, download, and delete Flight recordings.

The flight recording can be configure with predefined configurations and
user defined configurations that can be uploaded as *.jfc files.

Added WebTest for the new view.
Add a screenshot for the new view.