XIVY-12007 Add error column that shows number of failures and next retry
Moved columns so that they are in a more logical order

Add icon and links for channel and receiver.

Add column error that shows the total failed deliver attemps so far and
the next retry of the delivery. The tooltip shows also the last error
reason.

Move formating of durations to the new class DurationFormat.

Move code to create urls and icons for user and role to one place
Rename notification.xhtml  to notificationDeliveries.xhtml

Add test for DurationFormat.
Enhance test for notification deliveries.