XIVY-12827 Use equals method instead of comparing IDs to search nodes
In case a PermissionGroup and a Permission has by accident the same ID
the method searchPermissionNode might return the wrong node leading to a
ClassCastException if a group is returned instead of a permission.

Instead of using the ID use the equals method of Permission and
PermissionGroup to search for permission nodes. By doing so a Permission
is never equal to a PermissionGroup even if the IDs are the same.

Cherry-Pick 1ef7aba3dfa8537774a948dbeb4018f9f52c3ddb from master