XIVY-13513 Fix useSSL in ldap browser
If you configure the following in the Engine Cockpit for Microsoft Active Directory:
- URL: ldaps://test-ad.ivyteam.io:637
- User: admin@zugtstdomain.wan
- Password: ***
- DefaultContext: OU=IvyTeam Test-OU,DC=zugtstdomain,DC=wan
- EnableInsecureSSL: true

The Browser does not work. In Axon Ivy 10 you still need to set useSSL=true.
If you set new useSSL=true, it does still not work.

This is because the Engine Cockpit does not read this property correctly.
If this property isset, the IvySslSocketFactory comes into charge which is capable
to handle enableInsecure and is working with our own trust- and keystore.