XIVY-14130 Filter the correct bean to work on for poll and start/stop
Using the first bean in the list is not reliable as on some systems the
order is different an the first bean may not work correctly in case of
poll and start/stop.
Therefore, I now filter the correct bean to work on for testing poll and
start/stop.

Wait() throws a TimeoutException and not a UIAssertionException.
Therefore no screenshot is taken by default.
Adding the ScreenShooterExtension to tests were Wait() is used so that a
screenshot is taken if the test fails.