XIVY-14370 Display bean and process element name and description
The start and intermediate event views shows now the name and
description of the bean and also of the process element.
If no name is configured on the process element only the name of the
bean is displayed.
If the name is configured on the process element then the name of the
process element with the name of the bean in brackets as postfix is
displayed in the name column.
Same for the description column.