Fix tests that are broken because of a changed browser render pane size
Seems that a new version of the browser has a slightly smaller render
pane, which leads to broken tests because some UI elements are only
visible by some pixels but the clickable part is not. Selenide does not
scroll to the UI element because it is already visible.

Add an explicit scrollTo fixes the problem.