XIVY-15527 Improve the performance of the Applications view
Use the new API to get the number of running cases for a process model
that only fires one SQL statement and not one per PMV of the process
model.

Cache the deletable information so that it is not evaluated on every
request. It triggers also the running cases query.

Cherry-Pick 51b5d8ed5fb3a18bffe1f6a3f201043c9ebfe7c7 from master