XIVY-15527 Improve the performance of the Applications view
Use the new API to get the number of running cases for a process model
that only fires one SQL statement and not one per PMV of the process
model.

Cache the deletable information so that it is not evaluated on every
request. It triggers also the running cases query.

Cherry-Pick 2f0cc0a333b1b603b0b266083e6ac93d3e82b7e1 from master